requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-SSLify==0.1.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
olefile==0.46
Pillow==6.0.0
pyTelegramBotAPI==3.6.6
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
